refactor: Integrate GamePlayLayout and GameLayout
